:recycle: Refactor panel files into standard Vue Single File Component
* Refactor panel.html/css/js to standard Vue Single File Component
format
* Use JS ES6 import statements for referencing other components/libs.
This will help use bundlers and dependencies more effectively
* Implement SASS/SCSS and scoped styles for standardizing front-end
architecture
* Move re-usable button templates to a components directory
* WIP: Clean up old code once validating functionality
