+++ src/main/zapHomeFiles/hud/entrypoints/Panel/panel.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width,initial-scale=1.0">
+	<!-- <link rel="stylesheet" type="text/css" href="{{ ZAP_HUD_FILES }}?name=libraries/spectre.css"/> -->
+</head>
+<body>
+	<div id="app">
+	</div>
+
+	<script type="text/javascript" src="{{ ZAP_HUD_FILES }}?name=libraries/localforage.min.js"></script>
+	<script type="text/javascript" src="{{ ZAP_HUD_FILES }}?name=utils.js"></script>
+	<script src="./panel.js"></script>
+</body>
+</html>
