--- src/main/zapHomeFiles/hud/panel.css
@@ -1,3 +1,0 @@
-body {
-  background: none transparent;
-}
