--- src/main/zapHomeFiles/hud/panel.html
@@ -1,31 +1,0 @@
-<html>
-<head>
-	<link rel="stylesheet" type="text/css" href="{{ ZAP_HUD_FILES }}?name=libraries/spectre.css"/>
-	<link rel="stylesheet" type="text/css" href="{{ ZAP_HUD_FILES }}?name=panel.css&amp;orientation=ORIENTATION" />
-
-	<script type="text/javascript" src="{{ ZAP_HUD_FILES }}?name=libraries/localforage.min.js"></script>
-	<script type="text/javascript" src="{{ ZAP_HUD_FILES }}?name=libraries/vue.js"></script>
-	<script type="text/javascript" src="{{ ZAP_HUD_FILES }}?name=utils.js"></script>
-	<script type="text/javascript" src="{{ ZAP_HUD_FILES }}?name=panel.js"></script>
-</head>
-<body>
-	<div id="app">
-		<hud-buttons></hud-buttons>
-	</div>
-
-	<template id="hud-buttons-template">
-		<div id="hud-buttons" :class="{'d-hide': !isVisible, 'd-visible': isVisible}">
-			<hud-button v-for="tool in tools" :key="tool.name" :label="tool.label" :icon="'{{ ZAP_HUD_FILES }}?image=' + tool.icon" :data="tool.data" :name="tool.name"></hud-button>
-			<hud-button icon="{{ ZAP_HUD_FILES }}?image=plus.png" name="add-tool"></hud-button>
-		</div>
-	</template>
-
-	<template id="hud-button-template">
-		<div :class="{'hud-button': true, 'small': isSmall, 'active': isActive, 'disabled': isDisabled}" :style="{'float': orientation, 'direction': direction}" @click="selectButton" @contextmenu="showContextMenu" @mouseover="mouseOver" @mouseleave="mouseLeave">
-			<img :src="currentIcon" :style="{'float': orientation}">
-			<span id="data" v-show="showData" :style="{'float': orientation, 'margin-left': marginleft, 'margin-right': marginright}">{{currentData}}</span>
-			<span id="label" @transitionend="transitionEnd" :style="{'float': orientation, 'margin-left': labelmarginleft, 'margin-right': labelmarginright}">{{label}}</span>
-		</div>
-	</template>
-</body>
-</html>
